fix calcfunction source code
