+++ aiida_workgraph/utils/__init__.py
@@ -454,3 +454,84 @@ def recursive_to_dict(attr_dict):
         return {k: recursive_to_dict(v) for k, v in attr_dict.items()}
     else:
         return attr_dict
+
+
+def get_required_imports(func):
+    """Retrieve type hints and the corresponding module"""
+    from typing import get_type_hints, _SpecialForm
+
+    type_hints = get_type_hints(func)
+    imports = {}
+
+    def add_imports(type_hint):
+        if isinstance(
+            type_hint, _SpecialForm
+        ):  # Handle special forms like Any, Union, Optional
+            module_name = "typing"
+            type_name = type_hint._name or str(type_hint)
+        elif hasattr(
+            type_hint, "__origin__"
+        ):  # This checks for higher-order types like List, Dict
+            module_name = type_hint.__module__
+            type_name = type_hint._name
+            for arg in type_hint.__args__:
+                if arg is type(None):  # noqa: E721
+                    continue
+                add_imports(arg)  # Recursively add imports for each argument
+        elif hasattr(type_hint, "__module__"):
+            module_name = type_hint.__module__
+            type_name = type_hint.__name__
+        else:
+            return  # If no module or origin, we can't import it, e.g., for literals
+
+        if module_name not in imports:
+            imports[module_name] = set()
+        if type_name is not None:
+            imports[module_name].add(type_name)
+
+    for _, type_hint in type_hints.items():
+        add_imports(type_hint)
+
+    return imports
+
+
+def serialize_function(func: Callable) -> Dict[str, Any]:
+    """Serialize a function for storage or transmission."""
+    import inspect
+    import textwrap
+    import cloudpickle as pickle
+
+    # we need save the source code explicitly, because in the case of jupyter notebook,
+    # the source code is not saved in the pickle file
+    source_code = inspect.getsource(func)
+    # Split the source into lines for processing
+    source_code_lines = source_code.split("\n")
+    function_source_code = "\n".join(source_code_lines)
+    # Find the first line of the actual function definition
+    for i, line in enumerate(source_code_lines):
+        if line.strip().startswith("def "):
+            break
+    function_source_code_without_decorator = "\n".join(source_code_lines[i:])
+    function_source_code_without_decorator = textwrap.dedent(
+        function_source_code_without_decorator
+    )
+    # we also need to include the necessary imports for the types used in the type hints.
+    try:
+        required_imports = get_required_imports(func)
+    except Exception as e:
+        required_imports = {}
+        print(f"Failed to get required imports for function {func.__name__}: {e}")
+    # Generate import statements
+    import_statements = "\n".join(
+        f"from {module} import {', '.join(types)}"
+        for module, types in required_imports.items()
+    )
+    return {
+        "executor": pickle.dumps(func),
+        "type": "function",
+        "is_pickle": True,
+        "function_name": func.__name__,
+        "function_source_code": function_source_code,
+        "function_source_code_without_decorator": function_source_code_without_decorator,
+        "import_statements": import_statements,
+    }
